# storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/register": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "Register new user",
                "description": "Create new user",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Register new user data",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Register"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Register"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "Register"
                ],
                "summary": "login",
                "description": "login",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Register new user data",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Login"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Login"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/follow": {
            "get": {
                "tags": [
                    "Follow"
                ],
                "summary": "View all followed users",
                "description": "All followed users",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Tweet"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Follow"
                ],
                "summary": "Follow user using id",
                "description": "Store Follow",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        },
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "User that would be followed",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Follow"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/unfollow/{id}": {
            "delete": {
                "tags": [
                    "Follow"
                ],
                "summary": "Unfollow user using id",
                "description": "Store Follow",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        },
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Follow"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/tweets": {
            "get": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Get a listing of the Tweets.",
                "description": "Get all Tweets",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Follow"
                        },
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Tweet"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Store a newly created Tweet in storage",
                "description": "Store Tweet",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        },
                        "type": "string"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Tweet that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Tweet"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/tweets/{id}": {
            "get": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Display the specified Tweet",
                "description": "Get Tweet",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        },
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tweet",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Tweet"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Update the specified Tweet in storage",
                "description": "Update Tweet",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        },
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tweet",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Tweet that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Tweet"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Tweet"
                ],
                "summary": "Remove the specified Tweet from storage",
                "description": "Delete Tweet",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Tweet"
                        },
                        "type": "string"
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Tweet",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "Follow": {
            "required": [
                "followed_id"
            ],
            "properties": {
                "followed_id": {
                    "description": "followed_id",
                    "type": "integer",
                    "format": "int32"
                }
            }
        },
        "Tweet": {
            "required": [
                "text_en",
                "text_ar"
            ],
            "properties": {
                "text_en": {
                    "description": "text_en",
                    "type": "string"
                },
                "text_ar": {
                    "description": "text_ar",
                    "type": "string"
                }
            }
        },
        "Login": {
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "user email",
                    "type": "string",
                    "format": "email"
                },
                "password": {
                    "description": "password",
                    "type": "string",
                    "format": "password"
                }
            }
        },
        "Register": {
            "required": [
                "email",
                "name",
                "password",
                "confirm_password",
                "image"
            ],
            "properties": {
                "email": {
                    "description": "user email",
                    "type": "string",
                    "format": "email"
                },
                "name": {
                    "description": "user name",
                    "type": "string",
                    "format": "int32"
                },
                "password": {
                    "description": "password",
                    "type": "string",
                    "format": "password"
                },
                "confirm_password": {
                    "description": "password",
                    "type": "string",
                    "format": "password"
                },
                "image": {
                    "description": "password",
                    "type": "string",
                    "format": "uri"
                }
            }
        }
    }
}